# Optimize start up time by using rollup