--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@
 /node_modules/
 yarn-error.log
 yarn.lock
+dist/