--- a/.travis.yml
+++ b/.travis.yml
@@ -1,9 +1,6 @@
 sudo: false
 language: node_js
 node_js:
-  - 7
-  - 8
-  - 9
+  - node
+  - lts/*
   - 10
-  - 11
-  - stable