--- a/package.json
+++ b/package.json
@@ -10,7 +10,7 @@
     "test": "node test/basic.js"
   },
   "engines": {
-    "node": ">=7.0.0"
+    "node": ">=10.0.0"
   },
   "keywords": [
     "color",