Seeing chalk/chalk#300 I thought it's worth to try to see if rollup can do any good.

By bundling everything into one file and removing some redundant `Object.keys()` calls require time dropped by 1/3.

Old:

```
Start time: (2018-09-29 23:56:15 UTC) [treshold=0%]
 #  module                                                        time  %
 1  text-table (node_modules/text-table/index.js)                  1ms  ▇▇▇▇ 6%
 2  date-time (node_modules/date-time/index.js)                    1ms  ▇▇▇▇ 6%
 3  parse-ms (node_modules/parse-ms/index.js)                      2ms  ▇▇▇▇▇▇▇ 11%
 4  pretty-ms (node_modules/pretty-ms/index.js)                    3ms  ▇▇▇▇▇▇▇▇▇▇ 17%
 5  ansi-styles (node_modules/tim.../ansi-styles/ansi-styles.js)   1ms  ▇▇▇▇ 6%
 6  strip-ansi (node_modules/time...modules/strip-ansi/index.js)   1ms  ▇▇▇▇ 6%
 7  has-color (node_modules/has-color/index.js)                    2ms  ▇▇▇▇▇▇▇ 11%
 8  chalk (node_modules/time-requ...node_modules/chalk/index.js)   5ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 28%
 9  color-name (node_modules/color-name/index.js)                  1ms  ▇▇▇▇ 6%
10  ./conversions (conversions.js)                                 4ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 22%
11  ./conversions (conversions.js)                                 1ms  ▇▇▇▇ 6%
12  ./route (route.js)                                             1ms  ▇▇▇▇ 6%
13  ./index.js (index.js)                                          7ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 39%
Total require(): 13
Total time: 18ms
```

New:

```
Start time: (2018-09-29 23:56:55 UTC) [treshold=0%]
 #  module                                                        time  %
 1  text-table (node_modules/text-table/index.js)                  1ms  ▇▇▇▇ 6%
 2  date-time (node_modules/date-time/index.js)                    1ms  ▇▇▇▇ 6%
 3  parse-ms (node_modules/parse-ms/index.js)                      1ms  ▇▇▇▇ 6%
 4  pretty-ms (node_modules/pretty-ms/index.js)                    2ms  ▇▇▇▇▇▇▇ 11%
 5  ansi-styles (node_modules/tim.../ansi-styles/ansi-styles.js)   1ms  ▇▇▇▇ 6%
 6  strip-ansi (node_modules/time...modules/strip-ansi/index.js)   1ms  ▇▇▇▇ 6%
 7  has-color (node_modules/has-color/index.js)                    2ms  ▇▇▇▇▇▇▇ 11%
 8  chalk (node_modules/time-requ...node_modules/chalk/index.js)   6ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 33%
 9  color-name (node_modules/color-name/index.js)                  1ms  ▇▇▇▇ 6%
10  ./lib/index.js (lib/index.js)                                  5ms  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 28%
Total require(): 10
Total time: 18ms
```

Look at the bottom (everything after `chalk`) for the result for `color-convert`.

Tested using `time-require` and Node.js 10.10.

Note: This should be a semver major release as users can't deep-require into individual files anymore.