@Qix- 2018-10-01
> Build fails on Node.js < 6. Are you up for rising the limit?

Yep, < 6 are out of LTS. The codebase can (and should be) modernized. Just means it'll be a major release.