@realityking 2018-11-02
@Qix- I opened a PR to raise the requirement to Node 6 in #61

It'd be great if you cut a release of current master before merging #61 so we can have dome deduplication between chalk 2 and chalk 3. See this comment https://github.com/Qix-/color-convert/commit/c6b0e10c2463fda95c355b09433578d1e304999b#commitcomment-30760802