@Qix- 2018-12-06
Okay, that sounds fine.

I re-read the OP and realized my first two questions make little sense.

One more thing though: is there any way to get th BFS to run at build time? That would ultimately decrease include times in the most significant way.