@wtgtybhertgeghgtwtg 2018-12-06
You could probably get it to work with [`babel-plugin-preval`](https://github.com/kentcdodds/babel-plugin-preval) and some refactoring.  I can for a proof of concept if you want, but it'd probably be best to get some performance tests set up.