@Qix- 2018-12-06
Yeah a POC would be cool and appreciated. In terms of performance, I guarantee you the speedup would be worth it 😅.