@realityking 2018-12-11
[facebook/prepack](https://github.com/facebook/prepack) might be another option. But either way this PR is a good step into that direction. Could you merge it? 😁