@realityking 2019-01-08
For what it's worth, I did try it with prepack. That's a dead end as it currently doesn't allow CommonKJ modules as output.