@realityking 2019-01-31
Rebased for current master