@Qix- 2019-03-18
This doesn't really solve the startup time problems. External dependencies in this case are _tiny_. The problem with `color-convert` is that it does a BFS to find all routes. *That* is what needs to be done at compile time.

cc @sindresorhus