@Qix- 2019-03-18
```javascript
const fs = require('fs');
const path = require('path');

const cc = require('./conversions');
const route = require('./route');

const models = Object.keys(cc);

const gen = [
	'/* NOTE: AUTO-GENERATED SOURCE. DO NOT EDIT! */',
	fs.readFileSync(path.join(__dirname, 'conversions.js'), 'utf-8'),
	'\n/* NOTE: GENERATED ROUTES HERE! DO NOT EDIT! */'
];

function generate(conversion, raw) {
	conversion.reverse();
	if (raw) {
		conversion = conversion.map(c => `${c}.raw`);
	}
	return `(...args) => ${conversion.join('(')}(...args${')'.repeat(conversion.length)};`;
}

models.forEach(fromModel => {
	const routes = route(fromModel);
	const routeModels = Object.keys(routes);

	routeModels.forEach(toModel => {
		const conversion = routes[toModel].conversion;

		if (cc[conversion[0]][conversion[conversion.length - 1]]) {
			return;
		}

		gen.push(`convert['${fromModel}']['${toModel}'] = ${generate(conversion)}`);
	});
});

console.log(gen.join('\n'));
```

Here's is a rudimentary build script that resolves routes and prints out a script. However, it doesn't add the rounding functions.