@bfred-it 2019-03-28
I think `Proxy` would help here. Instead of finding all the conversion paths, just find them as needed and then "cache" the link. Something like:

```
proxy conversion; handler(from, to) ->
	if !conversion[from][to]
		conversion[from][to] = deriveBFSBetween (from, to)

	return conversion[from][to]
```