@Qix- 2019-03-28
I don't see why pre-processing them isn't a viable option?