@bfred-it 2019-03-28
Just a different approach. Preprocessing is good because the the code will run faster. Proxy is good for the developers because it doesn’t need a build step