@wtgtybhertgeghgtwtg 2019-03-28
Preprocessing doesn't make it run faster, it just makes it so there's less to do.  To a degree, proxying would work (and is kinda cool), but it'd have to do a weird dance to make sure it's still enumerable without evaluating everything immediately.