@bfred-it 2019-03-29
> Preprocessing doesn't make it run faster

Preprocessing would make the first run ("start up time") run faster and also every first-call after that, if you compare it to Proxy (which would have to run `deriveBFS`)