@Qix- 2019-03-29
Preprocessing isn't hard for developers. A build step isn't unjustifiable if it means doing an expensive step at compile time vs runtime. Also, enumeration would be impossible without running the bfs. Also, proxies are not free. They have been optimized heavily but they are not free.

I'm still in favor of a preprocessor step.