@wtgtybhertgeghgtwtg 2019-03-29
> I'm still in favor of a preprocessor step.

If you okay the method in #63, I can go ahead and rewrite that for `node@6` and rebase that.