@realityking 2020-09-27
I've updated the branch to be mergeable.

I know it's not the ultimate solution but since #63 is using rollup as well, I think it can be a nice stepping stone.