@realityking 2020-09-29
@sindresorhus All comments addressed