@Qix- 2021-02-28
Closing this. Thank you for the work and conversation everyone, but I'm going to implement an Acorn-based builder that calculates the paths and dumps a single script into a `dist` folder.

This will be part of the next major release.